move Simulation class out of simulation.__init__ into its own module
